review: Move ComponentBinder and relevant interfaces to flink-statefun-extensions
