+++ statefun-flink/statefun-flink-common/pom.xml
@@ -34,12 +34,6 @@ under the License.
     </properties>
 
     <dependencies>
-        <dependency>
-            <groupId>org.apache.flink</groupId>
-            <artifactId>statefun-sdk-embedded</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
         <!-- flink runtime -->
         <dependency>
             <groupId>org.apache.flink</groupId>
