+++ statefun-flink/statefun-flink-core/src/main/java/org/apache/flink/statefun/flink/core/httpfn/binders/HttpEndpointComponentBinderV2.java
@@ -21,9 +21,9 @@ package org.apache.flink.statefun.flink.core.httpfn.binders;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.DeserializationFeature;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.flink.statefun.flink.common.extensions.ComponentBinder;
-import org.apache.flink.statefun.flink.common.extensions.ExtensionResolver;
-import org.apache.flink.statefun.flink.common.json.ModuleComponent;
+import org.apache.flink.statefun.extensions.ComponentBinder;
+import org.apache.flink.statefun.extensions.ExtensionResolver;
+import org.apache.flink.statefun.extensions.ModuleComponent;
 import org.apache.flink.statefun.flink.core.httpfn.HttpFunctionEndpointSpec;
 import org.apache.flink.statefun.flink.core.httpfn.HttpFunctionProvider;
 import org.apache.flink.statefun.flink.core.httpfn.TargetFunctions;
