+++ statefun-flink/statefun-flink-io-bundle/src/test/java/org/apache/flink/statefun/flink/io/kinesis/binders/GenericKinesisEgressComponentBinderV1Test.java
@@ -7,8 +7,8 @@ import java.net.URL;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.JsonNode;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import org.apache.flink.statefun.flink.common.extensions.ExtensionResolver;
-import org.apache.flink.statefun.flink.common.json.ModuleComponent;
+import org.apache.flink.statefun.extensions.ExtensionResolver;
+import org.apache.flink.statefun.extensions.ModuleComponent;
 import org.apache.flink.statefun.flink.io.kafka.binders.AutoRoutableKafkaIngressComponentBinderV1Test;
 import org.apache.flink.statefun.flink.io.testutils.TestModuleBinder;
 import org.apache.flink.statefun.sdk.TypeName;
